+++ test/radix-sort/sorter.hpp
@@ -43,10 +43,10 @@ private:
     const size_t numPrefixScanGroups_;
 
     unsigned int *unsortedValues_;
-    const size_t numUnsortedElements_;
+    const unsigned long numUnsortedElements_; // size_t can't be passed to AMD
     const size_t numUnsortedBytes_;
     unsigned int *histogram_;
-    const size_t numHistogramElements_;
+    const unsigned long numHistogramElements_; // size_t can't be passed to AMD
     const size_t numHistogramBytes_;
 
     cl_mem unsortedValuesBuffer_;
